Move existing tournament home to old folder, create new component
