+++ src/app/tournaments/tournament-home/OLD/tournament-home.component.css
@@ -1,0 +1,25 @@
+.team-list {
+  justify-content: center;
+  align-items: center;
+}
+
+.team-card {
+  width: 240px;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.team-name {
+  font-weight: bold;
+  margin-right: 16px;
+}
+
+.no-rounds-warning {
+  width: 100%;
+  color: red;
+  text-align: center;
+  font-style: italic;
+}
