+++ src/app/tournaments/tournaments.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
 import { AngularMaterialModule } from '../angular-material.module';
-import { TournamentHomeComponent } from './tournament-home/tournament-home.component';
 import { TournamentStandingsComponent } from './tournament-standings/tournament-standings.component';
 import { TournamentScorecardComponent } from './tournament-scorecard/tournament-scorecard.component';
 import { ScorecardModule } from '../shared/scorecard/scorecard.module';
@@ -13,7 +12,6 @@ import { TournamentCreateComponent } from './tournament-create/tournament-create
 
 @NgModule({
   declarations: [
-    TournamentHomeComponent,
     TournamentStandingsComponent,
     TournamentScorecardComponent,
     TournamentScorecardCreateComponent,
